Update the text components to have more homogenized solutions for updating
