+++ Assets/Prefabs/UI/Adventure Narration UI.prefab
@@ -87,8 +87,9 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls:
         - m_Target: {fileID: 3889091749351599542}
-          m_TargetAssemblyTypeName: Game.GUI.Wiki.AdventureUIComponent, Assembly-CSharp
-          m_MethodName: ToggleCompleted
+          m_TargetAssemblyTypeName: Game.GUI.Wiki.AdventureUIComponent`1[[Game.Simulation.AdventureNarrationComponent,
+            Assembly-CSharp
+          m_MethodName: OnPointerClick
           m_Mode: 1
           m_Arguments:
             m_ObjectArgument: {fileID: 0}
