+++ Assets/Prefabs/UI/Adventure Text UI.prefab
@@ -261,7 +261,9 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_text: Text Goes Here --------------------------
+  m_text: 'Text Goes Here --------------------------
+
+    ddddddddddddddddddaaaaaaaaaaaa'
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
   m_sharedMaterial: {fileID: 2180264, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
