+++ Assets/Resources/ScriptableObjects/PopupDefinitions/CreatureInfoCardDefinition.asset
@@ -21,5 +21,5 @@ MonoBehaviour:
     PrefabModificationsReferencedUnityObjects: []
     PrefabModifications: []
     SerializationNodes: []
-  popupTypeString: CREATURE_INFO_CARD
-  prefab: {fileID: 0}
+  popupTypeString: MONSTER_INFO_CARD
+  prefab: {fileID: 5137877150577010650, guid: 0374efe10989169448231321fb5823ba, type: 3}
