+++ Assets/Scripts/Simulation/Adventures/IAdventureTextComponent.cs
@@ -1,0 +1,7 @@
+﻿namespace Game.Simulation
+{
+	public interface IAdventureTextComponent
+	{
+		public string Text { get; }
+	}
+}
