+++ Assets/Scripts/Simulation/Adventures/IAdventureTextComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d71ec4d23c3afe4f95408bc76dd7a27
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
